requirements.txt
aiosqlite==0.21.0
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
astor==0.8.1
asyncpg==0.30.0
bcrypt==4.0.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.2
decorator==5.2.1
deprecation==2.1.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.13
future==1.0.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
loguru==0.7.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
miniutils==1.0.3
multidict==6.7.0
orjson==3.11.3
packaging==25.0
passlib==1.7.4
pillow==11.3.0
pluggy==1.6.0
postgrest==2.22.0
pragma==0.2.5
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
PyContracts==1.8.12
pycparser==2.23
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pyreadline3==3.5.4
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
realtime==2.5.3
reportlab==4.4.4
rich==14.2.0
rich-toolkit==0.15.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.37.2
storage3==0.12.1
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
tqdm==4.67.1
typer==0.19.2
typing_extensions==4.15.0
ujson==5.11.0
uvicorn==0.23.2
watchfiles==1.1.0
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.22.0